requirements.txt
Django>=4.1,<4.2
djangorestframework>=3.13,<3.14
psycopg2>=2.9,<2.10
drf-spectacular<=0.24,<0.25
